requirements.txt
# Dia-Agent 依赖列表
# 核心依赖

# 向量数据库
chromadb>=0.4.0

# Embedding 和 Reranker
FlagEmbedding>=1.2.0
transformers>=4.30.0

# 关键词检索
rank-bm25>=0.2.2
jieba>=0.42.0

# 图数据库
neo4j>=5.0.0

# 数据处理
numpy>=1.24.0
pydantic>=2.0.0

# 可选：LLM 集成
# openai>=1.0.0
# langchain>=0.1.0
python-dotenv
